requirements.txt
betterproto==2.0.0b5
black==23.3.0
click==8.1.3
colorama==0.4.6
grpcio==1.54.2
grpcio-tools==1.54.2
grpclib==0.4.4
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
protobuf==4.23.1
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
